+++ example/App.js
@@ -53,10 +53,10 @@ export default class App extends Component {
           preferredBarTintColor: '#453AA4',
           preferredControlTintColor: 'white',
           readerMode: false,
-          animated: false,
+          animated: true,
           modalPresentationStyle: 'overFullScreen',
           modalTransitionStyle: 'partialCurl',
-          modalEnabled: false,
+          modalEnabled: true,
           // Android Properties
           showTitle: true,
           toolbarColor: '#6200EE',